@dhardy 2021-09-12
Thanks! Could you do `cargo +nightly bench --bench generators chacha` before/after?