@kazcw 2021-09-12
It is actually between 1-2% slower. Looking at the ASM, the optimizer does the right thing with the bytewise loop (unroll the loop, move data in SIMD chunks), but it doesn't see through the wordwise loop.

However, I found that if I manually unroll the loop, the optimizer produces SIMD output equivalent to the current `unsafe` version. While that adds 16 repetitious lines, it achieves safety without affecting current performance.

Benchmarks (SSE4.1 machine):

Before this PR:
test gen_bytes_chacha12      ... bench:   1,176,279 ns/iter (+/- 9,206) = 870 MB/s                                                                             
test gen_bytes_chacha20      ... bench:   1,774,785 ns/iter (+/- 17,849) = 576 MB/s                                                                            
test gen_bytes_chacha8       ... bench:     869,230 ns/iter (+/- 951) = 1178 MB/s

After original PR (the impact would be up to 2x as much on an AVX2 machine):
test gen_bytes_chacha12      ... bench:   1,193,299 ns/iter (+/- 9,527) = 858 MB/s
test gen_bytes_chacha20      ... bench:   1,809,063 ns/iter (+/- 1,675) = 566 MB/s
test gen_bytes_chacha8       ... bench:     892,924 ns/iter (+/- 2,077) = 1146 MB/s

After updated PR:
test gen_bytes_chacha12      ... bench:   1,165,595 ns/iter (+/- 9,184) = 878 MB/s
test gen_bytes_chacha20      ... bench:   1,773,808 ns/iter (+/- 2,155) = 578 MB/s
test gen_bytes_chacha8       ... bench:     870,023 ns/iter (+/- 1,885) = 1176 MB/s