@Ralith 2021-09-12
What if you use explicit indexing (`for i in 0..4`) rather than zipped iterators? I've found that tends to optimize down more reliably in hecs, at least. Failing that, seems worth leaving a comment to justify the hand-unroll to help provide context in the future.