@kazcw 2021-09-12
@Ralith that's the ticket! It's emitting SIMD without unrolling the loop, and it seems to be a tiny bit faster than the original.

Final version (0..4 loop):
test gen_bytes_chacha12      ... bench:   1,140,003 ns/iter (+/- 6,816) = 898 MB/s
test gen_bytes_chacha20      ... bench:   1,730,812 ns/iter (+/- 8,309) = 591 MB/s
test gen_bytes_chacha8       ... bench:     846,779 ns/iter (+/- 4,420) = 1209 MB/s