@kazcw 2021-09-12
@Ralith: Thanks for the idea, but I'm getting poor performance with a 0..4 loop. Tried as follows:

    for i in 0..4 {
        let j = i * 16;
        out[j..(j+4)].copy_from_slice(&(a[i] + k).to_lanes());
        out[(j+4)..(j+8)].copy_from_slice(&(b[i] + sb).to_lanes());
        out[(j+8)..(j+12)].copy_from_slice(&(c[i] + sc).to_lanes());
        out[(j+12)..(j+16)].copy_from_slice(&(d[i] + sd[i]).to_lanes());
    }