@vks 2021-09-15
> "Before" is ceb25f8 (master before merging this), "after" is 6e6b4ce.

So now is faster? Then what is the problem?